# Don't apply msvc link opts for non-opt build