@upsuper 2016-11-25
As a comparison, the build script of [libbindgen](https://github.com/servo/rust-bindgen/tree/master/libbindgen), with `/OPT:REF,ICF`, it takes 1m40s to link on my machine, but without that, 0.55s.