@retep998 2016-11-25
`gc_sections` for gnu linkers doesn't check whether optimizations are enabled before passing its equivalent flags. This would mean that we'd be ignoring `sess.opts.cg.link_dead_code` when we're using the msvc linker with optimizations disabled.

If we are going to disable `/OPT:REF,ICF`, we really ought to do incremental linking at the same time. https://github.com/rust-lang/rust/issues/37543