@upsuper 2016-11-27
I agree that we should do incremental linking as well. But I think that change could be more complicated than simply turning off the link-time optimization for debug build. And this trivial change shows a great win on link time, especially for crates who have lots of generated code.

As I've mentioned above, switching from `/OPT:REF,ICF` to nothing reduces link time for build script of libbindgen from ~100s to ~0.55s. Also it reduces the time of linking the whole [rust-bindgen](https://github.com/servo/rust-bindgen/) binary from ~185s to ~1.1s. It is a ~180x improvement.

So I think we should take it as soon as possible without letting further potential improvement block it.