@ollie27 2016-11-27
[`/INCREMENTAL`](https://msdn.microsoft.com/en-us/library/4khtbfyf.aspx) is enabled by default unless `/OPT` is also passed so this patch may be enough to enable incremental linking.