@retep998 2016-11-28
@ollie27 Cargo deletes the old binary before invoking rustc, so there is no chance of the result of previous linking being usable for incremental linking. Making incremental linking actually possible should be as simple as telling cargo to __not__ delete that sort of stuff, although it'll be a while before anyone actually teaches cargo when to do that.