@ollie27 2016-11-28
Well maybe we should explicitly pass `/INCREMENTAL:NO` until we fix Cargo so it doesn't generate unnecessary `.ilk` files etc.