@cpeterso 2016-11-30
@Manishearth or @pnkfelix, can you review @upsuper's linker change here? This is a big productivity issue. This PR reduces his link times from three minutes (!!) to one second by disabling MSVC LTO for debug builds.