@upsuper 2016-12-01
> Making incremental linking actually possible should be as simple as telling cargo to not delete that sort of stuff, although it'll be a while before anyone actually teaches cargo when to do that.

Incremental linking may need more investigation to see whether that works well with Rust's binaries. I did see an issue that incremental linked binary (linking a Rust library into Gecko) sometimes has weird runtime error inside Rust code.