@alexcrichton 2016-12-02
@upsuper 

Thanks for the PR! Sorry about taking a bit to get around to this. I'd be curious to play around a bit here with the various options before we just turn this off entirely. We get *significant* size wins sometimes with `--gc-sections`-equivalent options, so it's beneficial to have this turned on all the time if we can.

IIRC `REF` and `ICF` are separate optimizations. I don't recall ever measuring either independently, so I'd be curious about what the effect is by using only one of those rather than both. Could you measure the compile times there?

Also, could you take a look at the traditional "hello world" and see what the difference in the size of the executable is? Does it jump by megabytes or does it remain the same?

In any case I think we definitely need to change *something* here because a 100s link time is crazy huge. Just want to explore the possibility space ahead of time!