@upsuper 2016-12-03
I tried building the build script of libbindgen. This is the data:

OPT | Link time | Binary size
---- | ---------- | -----------
REF,ICF | 85.51s | 5,648 KB
REF | 82.17s | 5,648 KB
ICF | 143.89s | 7,790 KB
(None) | 0.71s | 12,795 KB

And for binary of rust-bindgen:

OPT | Link time | Binary size
---- | ---------- | -----------
REF,ICF | 147.81s | 9,283 KB
REF | 151.09s | 9,283 KB
ICF | 156.14s | 12,748 KB
(None) | 1.04s | 21,513 KB

I didn't run them many times, but it seems to me the time consumption is roughly stable, and somehow faster than my measurement before. Probably because that laptop is not doing anything else this time.

So it seems the binary size is indeed much larger without the optimizations (~2.3x), but the reduction of time consumption should well justify that difference.