@upsuper 2016-12-03
Oh, `/OPT:REF` indicates `/OPT:ICF` as well. Let me test again with `/OPT:REF,NOICF`.