@upsuper 2016-12-03
Updated the data with `/OPT:REF,NOICF`. So it seems `/OPT:REF,NOICF` might be a good compromise between size and time consumption, which is as fast as no optimization, but still reduces the size to some extent.