@upsuper 2016-12-03
Updated the patch to use `/OPT:REF,NOICF` for non-opt build.