@alexcrichton 2016-12-03
Thanks for the investigation @upsuper! Definitely makes sense to me to leave `REF` on and turn `ICF` off. 

@bors: r+